requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
python-telegram-bot[job-queue]
requests
